+++ odoo/local-src/specific_partner/models/res_partner.py
@@ -24,7 +24,6 @@ class ResPartner(models.Model):
     pharmacist_id = fields.Many2one(
         comodel_name='res.partner',
         string='Associated pharmacist',
-        domain=lambda self: self._get_pharmacist_id_domain()
     )
 
     @api.depends('alcyon_category_id')
@@ -34,14 +33,3 @@ class ResPartner(models.Model):
             partner.depot_number_visible = (
                 partner.alcyon_category_id == veterinary
             )
-
-    def _get_pharmacist_id_domain(self):
-        """ Filtering pharmacist partner on the associated alcyon category.
-        """
-        try:
-            category = self.env.ref('scenario.partner_category_pharmacy')
-        except ValueError:
-            domain = False
-        else:
-            domain = [('alcyon_category_id', '=', category.id)]
-        return domain